Adds search-results component and search service for YouTube API
I added a search-results component, which eventually will contain
a search-bar component and a search-display component. The search-bar
component will send a query parameter to the search service, which makes
a call to the YouTube API, returning an array of song objects,
which I will render to the search-display component (not yet
implemented). I also created a storage object using ember-local-storage
module, which I will use to store the YouTube API response, in order
to feed the results to the search-display component.